Store recent events under parent/child key
So we can, for instance, find a particular ReplicaSet event that was
actually reported on a Deployment.

We need some hacky parsing of the message string to find those, but
the rest is actually a bit cleaner than the previous
`recent`/`recentTopLevel` split.

Add a test to check that we report the correct parentage from events
which arrive in a less-ideal order.