Synthesise Events from Condition transitions
This gives a much richer trace output, showing how objects get
initialized.

We have to do a 'watch' on every object to see Conditions change over
time. We watch everything in the owner chain, to pick up changes at
every level.

We adjust timestamps from condition transitions that have arrived
promptly, the same as we do for Events, to give a more fine-grained
display in tracing.

Note the "Events" that are synthesised are just used as a convenient
internal object to turn into spans; they aren't meant to look like real
Events. For instance the ObjectMeta is populated with info useful for
debugging.