Tests: Cope with ordering of spans changing order
Apply a depth-first sort to the spans so they come out a predictable
order for checking against expected output.

Some of the expected output needs to be re-ordered to match, but by
eyeball there is nothing untoward here; the previous ordering was
arbitrary.

Also we need to do some more flushing at the end of each test.